backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
cryptography>=44.0.0
httpx>=0.28.0
python-multipart>=0.0.18
